WIP __sanitizer_struct_mallinfo vs 2 size assertion

+++ libsanitizer/sanitizer_common/sanitizer_platform_limits_posix.h
@@ -192,7 +192,7 @@ struct __sanitizer_struct_mallinfo {
 
 #if SANITIZER_LINUX && !SANITIZER_ANDROID
 struct __sanitizer_struct_mallinfo {
-  int v[10];
+  long v[10];
 };
 
 extern unsigned struct_ustat_sz;
